RetroDebugger cleanup

+++ Emulator/Utilities/IOUtils.h
@@ -95,10 +95,10 @@ struct dec {
 
 struct hex {
     
-    int digits;
+    isize digits;
     u64 value;
     
-    hex(int d, u64 v) : digits(d), value(v) { };
+    hex(isize d, u64 v) : digits(d), value(v) { };
     hex(u64 v) : hex(16, v) { };
     hex(u32 v) : hex(8, v) { };
     hex(u16 v) : hex(4, v) { };
@@ -108,10 +108,10 @@ struct hex {
 
 struct bin {
 
-    int digits;
+    isize digits;
     u64 value;
 
-    bin(int d, u64 v) : digits(d), value(v) { };
+    bin(isize d, u64 v) : digits(d), value(v) { };
     bin(u64 v) : bin(64, v) { };
     bin(u32 v) : bin(32, v) { };
     bin(u16 v) : bin(16, v) { };
@@ -130,10 +130,10 @@ struct flt {
 
 struct tab {
     
-    int pads;
+    isize pads;
     const string &str;
     
-    tab(int p, const string &s) : pads(p), str(s) { };
+    tab(isize p, const string &s) : pads(p), str(s) { };
     tab(const string &s) : tab(24, s) { };
     std::ostream &operator()(std::ostream &os) const;
 };
@@ -153,10 +153,10 @@ struct bol {
 
 struct str {
 
-    int characters;
+    isize characters;
     u64 value;
 
-    str(int c, u64 v) : characters(c), value(v) { };
+    str(isize c, u64 v) : characters(c), value(v) { };
     str(u64 v) : str(8, v) { };
     str(u32 v) : str(4, v) { };
     str(u16 v) : str(2, v) { };
